Added two files for RQ1
included the age_group in the candidates for the predictions and the leaderboard score, resulting in weirdly low scores and immens RAM usage, second file 'changed-...' is an attempt to tome in this usage and use functions. This is not yet finished and errors at the end because of the predictions generation that has not been changed yet.